requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
antiorm==1.2.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
awsebcli==3.15.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.4
blessed==1.15.0
blis==0.2.4
botocore==1.12.131
Bottleneck==1.2.1
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.3.9
cpplint==1.4.5
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.29.14
dask==2.1.0
db==0.1.1
decorator==4.4.2
defusedxml==0.6.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastai==1.0.50.post1
fastprogress==0.1.20
fastscript==0.1.4
filelock==3.0.12
fire==0.2.1
flufl.i18n==1.1.3
fn==0.4.3
future==0.16.0
gast==0.2.0
gdown==3.7.4
google-pasta==0.1.7
graphviz==0.13.2
grpcio==1.17.0
gym==0.10.11
h11==0.8.1
h5py==2.9.0
html5lib==0.9999999
httptools==0.0.13
idna==2.7
ig-cpp==0.1.9
imageio==2.5.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.4
kiwisolver==1.0.1
lxml==4.3.4
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
multidict==4.5.2
murmurhash==1.0.2
nbconvert==5.6.1
nbdev==0.2.13
nbformat==5.0.4
networkx==2.3
nose==1.3.7
notebook==6.0.3
numexpr==2.6.9
numpy==1.18.2
nvidia-ml-py3==7.352.0
onnx==1.4.1
open3d==0.9.0.0
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.6.2
Pat==0.5.2
path.py==11.5.0
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.2.0
plac==0.9.6
plotly==4.3.0
plyfile==0.7.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.10.0
ptyprocess==0.6.0
pyamplitude==1.2.0.dev1
pycodestyle==2.5.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.6.1
pymesh==1.0.2
pyntcloud==0.1.2
pynvx==1.0.0
pyobjc==6.1
pyparsing==2.3.1
pyperclip==1.7.0
PyQt5==5.13.2
PyQt5-sip==4.19.19
pyrosbag==0.1.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-multipart==0.0.5
python-pcl==0.3.0a1
python-telegram-bot==11.1.0
pytz==2018.9
PyWavelets==1.0.3
PyYAML==3.13
pyzmq==19.0.0
qtconsole==4.5.5
requests==2.20.1
retrying==1.3.3
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.0
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==1.9
spacy==2.1.3
spotlight==0.1.6
srsly==0.0.5
starlette==0.11.4
tarjan==0.2.3.2
TBB==0.1
telegram==0.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==0.9.1
thinc==7.0.4
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
torchviz==0.0.1
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.24.1
uvicorn==0.6.1
uvloop==0.12.2
virtualenv==16.4.3
wasabi==0.2.1
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.56.0
websockets==7.0
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
yarl==1.3.0
zipp==3.1.0
